requirements.txt
streamlit
google_generativeai
python-dotenv
langchain
PyPDF2
chormadb
Pillow
